fix: Navigation to the forum post and the explorer
